refactor: move packument fetch
Move logic for fetching packument from a registry into the io folder since it basically only abstracts an interaction with the outside world without adding any business logic. It is not really a service function
