+++ src/services/dependency-resolving.ts
@@ -18,9 +18,9 @@ import {
   ResolveLatestVersionService,
 } from "./resolve-latest-version";
 import { Err, Ok, Result } from "ts-results-es";
-import { FetchPackumentError } from "./fetch-packument";
 import { PackumentNotFoundError } from "../common-errors";
 import { HttpErrorBase } from "npm-registry-fetch/lib/errors";
+import {FetchPackumentError} from "../io/packument-io";
 
 export type DependencyBase = {
   /**
